aggressively clean up desktop with new inline poll
