+++ packages/desktop/Cargo.toml
@@ -35,6 +35,7 @@ dunce = "1.0.2"
 
 interprocess = { version = "1.1.1", optional = true}
 futures-util = "0.3.25"
+sledgehammer = "0.2.0"
 
 [target.'cfg(target_os = "ios")'.dependencies]
 objc = "0.2.7"
